Перенесла методы в helpers и поправила тесты в соответсвии с изменениями
